Backend/price_optimization_tool/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
configparser==7.2.0
flasgger==0.9.7.1
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
six==1.17.0
SQLAlchemy==2.0.42
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
